Albertsons Trucking Demo/mcp-server/requirements.txt
mcp[cli]>=1.6.0,<2.0.0
httpx==0.27.2
azure-cosmos==4.7.0
azure-identity==1.19.0
uvicorn[standard]==0.32.0
starlette==0.41.3
